# api/openapi.yaml
openapi: 3.0.3
info:
  title: AntiNAT Controller API
  version: 1.0.0-beta
  description: >
    Frozen AntiNAT v1.0-beta management API. All resource modifications use
    ETag/If-Match; missing precondition returns 428, revision mismatch 412,
    and port/state conflict 409. Creates accept a durable Idempotency-Key.
    Error responses use the registry in docs/error-codes.md.
servers:
  - url: /
...
paths:
...
  /api/v1/nodes:
    get:
      tags: [nodes]
      operationId: listNodes
      summary: List nodes
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/PageParam" }
        - { $ref: "#/components/parameters/PageSizeParam" }
        - { $ref: "#/components/parameters/SortParam" }
        - { $ref: "#/components/parameters/FilterParam" }
      responses:
        "200":
          description: Page of nodes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodePage" }
    post:
      tags: [nodes]
      operationId: createNode
      summary: Create a node
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKeyHeader" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NodeCreate" }
      responses:
        "201":
          description: Node created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        "400":
          description: Malformed request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Idempotency-key conflict or duplicate
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          description: Request body too large
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Per-principal rate limit exceeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: Unexpected server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/nodes/{id}/deployment-profile:
    parameters:
      - { $ref: "#/components/parameters/NodeIdParam" }
    get:
      tags: [deployment]
      operationId: getDeploymentProfile
      summary: Read a structured, secret-free deployment profile
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Deployment profile and current revision
          headers:
            Cache-Control:
              schema: { type: string, enum: ["private, no-store"] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeploymentProfileView" }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Node not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    put:
      tags: [deployment]
      operationId: saveDeploymentProfile
      summary: Save a structured deployment profile without a token or command
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/IfMatchHeader" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeploymentProfileRequest" }
      responses:
        "200":
          description: Profile saved with the next revision
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeploymentProfileView" }
        "400":
          description: Malformed or unknown-field request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "412":
          description: ETag/revision mismatch
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Semantically invalid profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "428":
          description: ETag/If-Match required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/forwards:
    get:
      tags: [forwards]
      operationId: listForwards
      summary: List forwards
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/PageParam" }
        - { $ref: "#/components/parameters/PageSizeParam" }
        - { $ref: "#/components/parameters/SortParam" }
        - { $ref: "#/components/parameters/FilterParam" }
      responses:
        "200":
          description: Page of forwards
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForwardPage" }
    post:
      tags: [forwards]
      operationId: createForward
      summary: Create a forward
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKeyHeader" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForwardCreate" }
      responses:
        "201":
          description: Forward created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Forward" }
        "409":
          description: Port/state conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/audit:
    get:
      tags: [observability]
      operationId: listAudit
      summary: List admin audit log entries
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/PageParam" }
        - { $ref: "#/components/parameters/PageSizeParam" }
        - { $ref: "#/components/parameters/CursorParam" }
      responses:
        "200":
          description: Page of audit entries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditPage" }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: antinat_session
  parameters:
    PageParam:
      name: page
      in: query
      description: One-based page number
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      description: Page size, 1..200
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    SortParam:
      name: sort
      in: query
      description: Sort field and direction, e.g. -created_at
      schema: { type: string }
    FilterParam:
      name: filter
      in: query
      description: Field filter, e.g. state=ONLINE
      schema: { type: string }
    CursorParam:
      name: cursor
      in: query
      description: Opaque pagination cursor (preferred over page for streams)
      schema: { type: string }
    IfMatchHeader:
      name: If-Match
      in: header
      required: true
      description: ETag precondition; missing returns 428
      schema: { type: string }
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Durable idempotency key for creates
      schema: { type: string, minLength: 8, maxLength: 128 }
...
    NodeIdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string, description: Error code from docs/error-codes.md }
        message: { type: string }
        request_id: { type: string }
        details:
          type: object
          additionalProperties: true
...
    Node:
      type: object
      required: [id, name, control_state, created_at, etag]
      properties:
        id: { type: string }
        name: { type: string }
        control_state: { type: string, enum: [ONLINE, OFFLINE] }
        os_arch: { type: string }
        network_fingerprint: { type: string }
        created_at: { type: string, format: date-time }
        etag: { type: string }
    NodeCreate:
      type: object
      properties:
        name: { type: string, description: Optional; generated when empty }
...
    NodePage:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Node" }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
        next_cursor: { type: string }
...
    DeploymentProfile:
      type: object
      additionalProperties: false
      required: [platform, controller_endpoint, detection_scheduler, log_level, auto_update]
      properties:
        platform: { type: string, enum: [linux, windows, docker] }
        controller_endpoint: { type: string, format: uri, description: HTTP(S) endpoint without credentials/query/fragment }
        bind_interface: { type: string }
        detection_scheduler: { type: string, enum: [sequential, parallel] }
        github_proxy: { type: string, format: uri }
        install_dir: { type: string }
        service_name: { type: string }
        log_level: { type: string, enum: [debug, info, warn, error] }
        auto_update: { type: string, enum: [disabled, manual, stable, enabled] }
    DeploymentProfileRequest:
      type: object
      additionalProperties: false
      required: [profile]
      properties:
        profile: { $ref: "#/components/schemas/DeploymentProfile" }
    DeploymentProfileView:
      type: object
      additionalProperties: false
      required: [node_id, profile, revision, etag]
      properties:
        node_id: { type: string }
        profile: { $ref: "#/components/schemas/DeploymentProfile" }
        revision: { type: integer, format: int64, minimum: 0 }
        etag: { type: string }
...
    Forward:
      type: object
      required: [id, node_id, name, protocol, target, strategy, desired_revision, etag]
      properties:
        id: { type: string }
        node_id: { type: string }
        name: { type: string }
        protocol: { type: string, enum: [tcp, udp] }
        target: { type: string, description: IPv4:port }
        strategy: { type: string }
        source_interface: { type: string }
        requested_local_port: { type: integer }
        requested_public_port: { type: integer }
        manual_expected_endpoint: { type: string }
        rate_limit_bps: { type: integer }
        detailed_stats: { type: boolean }
        publish_scheme: { type: string, enum: [http, https] }
        published_host: { type: string }
        custom_uri_template: { type: string }
        desired_revision: { type: integer }
        etag: { type: string }
        states:
          $ref: "#/components/schemas/ActivationStates"
    ActivationStates:
      type: object
      properties:
        control_state: { type: string }
        listener_state: { type: string }
        mapping_state: { type: string }
        keepalive_state: { type: string }
        wan_reachability_state: { type: string }
        return_path_state: { type: string }
        target_health_state: { type: string }
        publication_state: { type: string }
        data_plane_state: { type: string }
    ForwardCreate:
      type: object
      required: [node_id, name, protocol, target, strategy]
      properties:
        node_id: { type: string }
        name: { type: string }
        protocol: { type: string, enum: [tcp, udp] }
        target: { type: string }
        strategy: { type: string }
        source_interface: { type: string }
        requested_local_port: { type: integer }
        requested_public_port: { type: integer }
        manual_expected_endpoint: { type: string }
        rate_limit_bps: { type: integer }
        detailed_stats: { type: boolean }
        publish_scheme: { type: string, enum: [http, https] }
        published_host: { type: string }
        custom_uri_template: { type: string }
...
    ForwardPage:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Forward" }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
        next_cursor: { type: string }
...
    AuditPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AuditEntry" }
        next_cursor: { type: string }
    AuditEntry:
      type: object
      required: [id, event, created_at]
      properties:
        id: { type: string }
        event: { type: string }
        detail: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
...